+++ R/define_method.R
@@ -106,7 +106,7 @@ packages_search_native <- function(packages,
                                    ),
                                    platform = NULL,
                                    additional_channels = NULL,
-                                   verbose = FALSE) {
+                                   verbose = "silent") {
   channels_arg <- format_channels_args(
     additional_channels,
     channels