+++ R/utils.R
@@ -72,9 +72,10 @@ env_exists <- function(env_name = "condathis-env") {
   }
 }
 
-#' Create Package Base Env
+#' Create Package Base Environment
+#' @keywords internal
 create_base_env <- function() {
   if (isFALSE(env_exists(env_name = "condathis-env"))) {
-    create_env(env_name = "condathis-env", verbose = FALSE)
+    create_env(env_name = "condathis-env", verbose = "silent")
   }
 }